# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: DaYu Platform HTTP API
  version: 1.0.0-draft
  description: |
    大禹平台第一版 HTTP 契约。所有业务时间均为 UTC ISO-8601。
    浏览器先通过 GET /api/v1/session 获取与当前 Session 绑定的 CSRF Token；
    所有 POST、PUT、PATCH、DELETE 请求都必须携带 X-CSRF-TOKEN。
    本契约不暴露服务器绝对路径、NC 相对路径或 Nginx 内部下载位置。
servers:
  - url: /
...
paths:
...
  /api/v1/downloads:
    post:
      tags: [Download]
      summary: 登记用途并申请短期 NC 下载授权
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DownloadRequest'}
      responses:
        '201':
          description: 授权成功；downloadUrl 不包含服务器路径
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DownloadAuthorization'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/AssetGone'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
  /api/v1/admin/dashboard:
    get:
      tags: [Operations]
      summary: 查询管理仪表盘
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/FromOptional'
        - $ref: '#/components/parameters/ToOptional'
        - {name: productCode, in: query, schema: {$ref: '#/components/schemas/ProductCode'}}
        - {name: dataMode, in: query, schema: {$ref: '#/components/schemas/DataMode'}}
      responses:
        '200':
          description: 默认最近 30 天，最大范围 365 天
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Dashboard'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
  /api/v1/admin/index-scans:
    post:
      tags: [Operations]
      summary: 异步触发一次手动增量扫描
      security: [{sessionCookie: [], csrfToken: []}]
      responses:
        '202':
          description: 扫描任务已接受
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScanRunSummary'}
        '409':
          description: 已有扫描运行
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    get:
      tags: [Operations]
      summary: 分页查询扫描历史
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/FromOptional'
        - $ref: '#/components/parameters/ToOptional'
        - {name: trigger, in: query, schema: {$ref: '#/components/schemas/ScanTrigger'}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/ScanStatus'}}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 扫描历史
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScanRunPage'}
  /api/v1/admin/index-scans/{scanRunId}:
    get:
      tags: [Operations]
      summary: 查询扫描任务详情和安全错误摘要
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/ScanRunIdPath'}]
      responses:
        '200':
          description: 扫描详情
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScanRunDetail'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/admin/download-audits:
    get:
      tags: [Operations]
      summary: 分页查询下载审计
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/FromOptional'
        - $ref: '#/components/parameters/ToOptional'
        - {name: productCode, in: query, schema: {$ref: '#/components/schemas/ProductCode'}}
        - {name: userId, in: query, schema: {type: integer, format: int64}}
        - {name: organization, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/DownloadStatus'}}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 下载审计分页；不返回 IP、User-Agent 或路径
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DownloadAuditPage'}
  /api/v1/admin/download-statistics:
    get:
      tags: [Operations]
      summary: 查询授权下载统计
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/FromOptional'
        - $ref: '#/components/parameters/ToOptional'
        - {name: productCode, in: query, schema: {$ref: '#/components/schemas/ProductCode'}}
      responses:
        '200':
          description: 统计只声称申请或授权，不声称客户端已保存
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DownloadStatistics'}
  /api/v1/admin/users:
    get:
      tags: [Operations]
      summary: 分页查询用户
      security: [{sessionCookie: []}]
      parameters:
        - {name: email, in: query, schema: {type: string}}
        - {name: organization, in: query, schema: {type: string}}
        - {name: role, in: query, schema: {$ref: '#/components/schemas/UserRole'}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/UserStatus'}}
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 用户分页，不返回密码哈希
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserPage'}
  /api/v1/admin/users/{userId}/status:
    put:
      tags: [Operations]
      summary: 修改用户状态
      security: [{sessionCookie: [], csrfToken: []}]
      parameters: [{$ref: '#/components/parameters/UserIdPath'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: {$ref: '#/components/schemas/UserStatus'}
      responses:
        '200':
          description: 更新后的用户
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserSummary'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/admin/users/{userId}/role:
    put:
      tags: [Operations]
      summary: 修改用户角色
      security: [{sessionCookie: [], csrfToken: []}]
      parameters: [{$ref: '#/components/parameters/UserIdPath'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: {$ref: '#/components/schemas/UserRole'}
      responses:
        '200':
          description: 更新后的用户
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserSummary'}
        '409': {$ref: '#/components/responses/Conflict'}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: DAYUSESSID
      description: HttpOnly、SameSite=Lax；生产 HTTPS 下 Secure=true
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: 必须与当前匿名或认证 Session 匹配
  parameters:
...
    UserIdPath:
      name: userId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
    ScanRunIdPath:
      name: scanRunId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
    FromOptional:
      name: from
      in: query
      schema: {$ref: '#/components/schemas/UtcInstant'}
    ToOptional:
      name: to
      in: query
      schema: {$ref: '#/components/schemas/UtcInstant'}
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: pageSize
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    Unauthenticated:
      description: 未登录或认证已失效
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Forbidden:
      description: 无权限、CSRF 失败或路径安全校验失败
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: 产品、资产或事件不存在；公开接口也用它隐藏未发布产品
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    AssetGone:
      description: 已索引资产对应文件丢失或短期授权过期
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Conflict:
      description: 唯一性、状态机或并发任务冲突
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    ValidationFailed:
      description: 参数格式正确但不满足业务校验
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
    ProductCode: {type: string, pattern: '^[A-Z][A-Z0-9_]{1,63}$'}
    UtcInstant: {type: string, format: date-time, pattern: 'Z$'}
    DataMode: {type: string, enum: [REALTIME, FORECAST]}
...
    UserRole: {type: string, enum: [USER, ADMIN]}
    UserStatus: {type: string, enum: [ACTIVE, DISABLED]}
    DownloadStatus: {type: string, enum: [REQUESTED, AUTHORIZED, DENIED, DELIVERED, INTERRUPTED]}
    ScanTrigger: {type: string, enum: [STARTUP, SCHEDULED, MANUAL]}
    ScanStatus: {type: string, enum: [RUNNING, SUCCEEDED, PARTIAL, FAILED]}
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: {type: string}
        message: {type: string}
        traceId: {type: string}
        details: {type: object, additionalProperties: true}
...
    DownloadRequest:
      type: object
      required: [assetId, purpose]
      properties:
        assetId: {type: integer, format: int64, minimum: 1}
        purpose: {type: string, minLength: 10, maxLength: 2000}
    DownloadAuthorization:
      type: object
      required: [downloadEventId, downloadUrl, fileName, expectedBytes, expiresAt]
      properties:
        downloadEventId: {type: integer, format: int64}
        downloadUrl: {type: string, format: uri-reference}
        fileName: {type: string}
        expectedBytes: {type: integer, format: int64, minimum: 0}
        expiresAt: {$ref: '#/components/schemas/UtcInstant'}
...
    ScanRunSummary:
      type: object
      required: [scanRunId, trigger, status, startedAt]
      properties:
        scanRunId: {type: integer, format: int64}
        trigger: {$ref: '#/components/schemas/ScanTrigger'}
        status: {$ref: '#/components/schemas/ScanStatus'}
        startedAt: {$ref: '#/components/schemas/UtcInstant'}
        finishedAt: {oneOf: [{$ref: '#/components/schemas/UtcInstant'}, {type: 'null'}]}
        scannedFiles: {type: integer, format: int64, minimum: 0}
        createdAssets: {type: integer, format: int64, minimum: 0}
        updatedAssets: {type: integer, format: int64, minimum: 0}
        removedWebpAssets: {type: integer, format: int64, minimum: 0}
        missingNetcdfAssets: {type: integer, format: int64, minimum: 0}
        errorCount: {type: integer, minimum: 0}
    ScanError:
      type: object
      required: [relativePath, errorCode, safeMessage]
      properties:
        relativePath: {type: string}
        errorCode: {type: string}
        safeMessage: {type: string}
    ScanRunDetail:
      allOf:
        - $ref: '#/components/schemas/ScanRunSummary'
        - type: object
          required: [errors]
          properties:
            errors: {type: array, items: {$ref: '#/components/schemas/ScanError'}}
    ScanRunPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/ScanRunSummary'}}
        page: {type: integer}
        pageSize: {type: integer}
        total: {type: integer, format: int64}
...
    DownloadStatistics:
      type: object
      required: [authorizedRequests, uniqueUsers, uniqueOrganizations, uniqueAssets]
      properties:
        authorizedRequests: {type: integer, format: int64, minimum: 0}
        uniqueUsers: {type: integer, format: int64, minimum: 0}
        uniqueOrganizations: {type: integer, format: int64, minimum: 0}
        uniqueAssets: {type: integer, format: int64, minimum: 0}
    Dashboard:
      type: object
      required: [publishedProducts, previewAvailableProducts, downloadAvailableProducts, missingProducts, staleProducts, downloads]
      properties:
        publishedProducts: {type: integer, format: int64}
        previewAvailableProducts: {type: integer, format: int64}
        downloadAvailableProducts: {type: integer, format: int64}
        missingProducts: {type: integer, format: int64}
        staleProducts: {type: integer, format: int64}
        latestScan: {oneOf: [{$ref: '#/components/schemas/ScanRunSummary'}, {type: 'null'}]}
        downloads: {$ref: '#/components/schemas/DownloadStatistics'}
    DownloadAudit:
      type: object
      required: [downloadEventId, userId, organizationSnapshot, assetId, productSnapshot, fileName, expectedBytes, purpose, authorizedAt, status]
      properties:
        downloadEventId: {type: integer, format: int64}
        userId: {type: integer, format: int64}
        organizationSnapshot: {type: string}
        assetId: {type: integer, format: int64}
        productSnapshot: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/ProductCode'}}
        fileName: {type: string}
        expectedBytes: {type: integer, format: int64}
        purpose: {type: string}
        authorizedAt: {$ref: '#/components/schemas/UtcInstant'}
        status: {$ref: '#/components/schemas/DownloadStatus'}
    DownloadAuditPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/DownloadAudit'}}
        page: {type: integer}
        pageSize: {type: integer}
        total: {type: integer, format: int64}
    UserSummary:
      type: object
      required: [userId, email, organization, role, status]
      properties:
        userId: {type: integer, format: int64}
        email: {type: string, format: email}
        organization: {type: string}
        role: {$ref: '#/components/schemas/UserRole'}
        status: {$ref: '#/components/schemas/UserStatus'}
    UserPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/UserSummary'}}
        page: {type: integer}
        pageSize: {type: integer}
        total: {type: integer, format: int64}